✨ コメントもらった部分を修正&expand_ecmds関数の位置と引数を変更

+++ src/data/new_struct_text_list.c
@@ -6,7 +6,7 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 19:56:50 by miyuu             #+#    #+#             */
-/*   Updated: 2025/03/03 20:26:03 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/03/04 01:35:20 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
  * ----------------------------
  * Creates a new t_text_list node with a str
  */
-t_text_list	*new_struct_text_list(char *str, int len)
+t_text_list	*new_struct_text_list(char *str, size_t len)
 {
 	t_text_list	*node;
